+++ src/components/hello-world-text.js
@@ -1,0 +1,9 @@
+import { Text } from 'react-native';
+
+const HelloWorldText = () => {
+  return (
+    <Text testID="hello-world-text">Hello World!!!</Text>
+  );
+};
+
+export default HelloWorldText;