+++ tst/utils/test-constants.js
@@ -1,0 +1,5 @@
+const HELLO_WORLD_TEXT = 'Hello World!!!';
+
+module.exports = {
+    HELLO_WORLD_TEXT,
+};